refactor: Improve performance 2
